Move custom log to util
